docs: Readme refresh (#17246)
## Description

<!-- Provide a 1-2 sentence description of your change -->

The README needed some love to show the current state of this project.

Updated quite a few links, removed some outdated information, and to a
larger extent now refers to MS Learn links. Most of the previous
information is now covered in-depth and actively updated on Learn.

Also added some info about social media and community which I think is
relevant for all the users! I didn't find anything about this in Learn
so figured this was a good place to put it.

Did a minor adjustment to the goals section regarding extensibility.
Didn't want to make big changes here, but it might be time with a
re-iteration of the goals (and non-goals) soon? :slightly_smiling_face:

Feel free to add to this if needed!

## Checklist

- [x] I have read and adhere to the [contribution
guide](https://github.com/Azure/bicep/blob/main/CONTRIBUTING.md).

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/17246)

---------

Co-authored-by: Anthony Martin <38542602+anthony-c-martin@users.noreply.github.com>