Type Builder Refactor and Unit test fix
Updated Type definition builder to have a simpler constructor. Initially
Type? was injected -- to fix this anti-pattern it is now wrapped inside
an options class with other members. This fixes potential ambiguity with
DI as well. Updated unit tests to consider these changes and added more
test cases to cover a wide area of scenarios.